{"schema":"decodex/commit/1","summary":"Modularize operator HTTP and dashboard assets","authority":"XY-1115"}
